+++ CHANGELOG.md
@@ -5,11 +5,39 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 3.1.0 - 2023-06-17
+
+### Added
+
+* `functions.php` has been added and included in the autoload to make wrapper function available globally (#19)
+  * `dotGet()` has been added to wrap `Dot::Get()` static method (#19)
+  * `dotSet()` has been added to wrap `Dot::Set()` static method (#19)
+  * `dotHas()` has been added to wrap `Dot::Has()` static method (#19)
+  * `dotCount()` has been added to wrap `Dot::Count()` static method (#19)
+  * `dotDelete()` has been added to wrap `Dot::Delete()` static method (#19)
+  * `dotAppend()` has been added to wrap `Dot::Append()` static method (#19)
+  * `dotFlatten()` has been added to wrap `Dot::Flatten()` static method (#19)
+
+## 3.0.0 - 2023-06-16
+
+### Added
 
 * There will be a 3.x branch created for php 8.0, there is little differences between the functionality of the different
   version of the package. The primary difference is the supported versions which have to do with the development of the
   package and the dev dependencies.
+
+## 2.1.0 - 2023-06-17
+
+### Added
+
+* `functions.php` has been added and included in the autoload to make wrapper function available globally (#19)
+  * `dotGet()` has been added to wrap `Dot::Get()` static method (#19)
+  * `dotSet()` has been added to wrap `Dot::Set()` static method (#19)
+  * `dotHas()` has been added to wrap `Dot::Has()` static method (#19)
+  * `dotCount()` has been added to wrap `Dot::Count()` static method (#19)
+  * `dotDelete()` has been added to wrap `Dot::Delete()` static method (#19)
+  * `dotAppend()` has been added to wrap `Dot::Append()` static method (#19)
+  * `dotFlatten()` has been added to wrap `Dot::Flatten()` static method (#19)
 
 ## 2.0.0 - 2023-06-15